fix: eliminate all compiler warnings from the build
Move C-linkage declarations into aura_jit_bridge.h and fiber.h so
module GMF partitions include headers instead of bare extern "C".
Fix FlatAST ctor initializer order, JIT fallback counter definition,
hash-table narrowing cast, and test missing-field/format warnings.

+++ src/compiler/evaluator_primitives_vector.cpp
@@ -343,7 +343,7 @@ void register_vector_and_hash_primitives(
                 pairs.push_back({EvalValue{keys[i - 1]}, EvalValue{vals[i - 1]}});
                 // Prepend to the result list.
                 auto list_pid = pairs.size();
-                pairs.push_back({EvalValue{kv_pid}, result});
+                pairs.push_back({EvalValue{static_cast<std::int64_t>(kv_pid)}, result});
                 result = make_pair(list_pid);
             }
         }
